+++ crates/rbt/src/lib.rs
@@ -13,7 +13,7 @@
 //! **Library**
 //! ```toml
 //! [dependencies]
-//! rbt-datalake = "0.3.6"
+//! rbt-datalake = "0.3.7"
 //! ```
 //!
 //! ## Quick start (library)
@@ -30,7 +30,7 @@
 //! # }
 //! ```
 
-#![doc(html_root_url = "https://docs.rs/rbt-datalake/0.3.6")]
+#![doc(html_root_url = "https://docs.rs/rbt-datalake/0.3.7")]
 
 pub mod core;
 pub mod engine;